add VerifiedDictionary
