+++ src/org/usfirst/frc/team5002/io/pipeline/stages/PipelineStagesFactory.java
@@ -24,4 +24,8 @@ package org.usfirst.frc.team5002.io.pipeline.stages;
 
 public class PipelineStagesFactory {
 
+    private PipelineStagesFactory() {}
+    public static void init() {
+
+    }
 }
